chore: separate core audio to dedicated source file and make initialisation from lua rather than being tightly coupled to video
